arm64: morello: Feature detection and basic system setup
This patch adds runtime detection for Morello, exposing
HWCAP2_MORELLO when detected, and just enough setup to allow Morello
instructions to be used without trapping. Capabilities cannot be loaded
or stored in memory yet.

The feature needs to be detected very early on the boot CPU, when
IRQs are still disabled, because disabling trapping of Morello
instructions implicitly enables capability exception entry and
return, requiring CVBAR_EL1 to be configured before any exception
can be taken, and CELR_EL1 when returning from an exception.

Morello relies on and mandates ARMv8.2-TTPBHA being implemented, as
access to capabilities in memory is controlled by fields allocated
in the PBHA bits. If the hardware advertises support for Morello but
not TTPBHA, something went wrong and we treat Morello as not
implemented.

Signed-off-by: Kevin Brodsky <kevin.brodsky@arm.com>