arm64: morello: Save root capability on boot
We will soon need to initialise userspace capabilities from a root
capability. Ensure this capability remains constant by copying it
from DDC to a global. We also warn if the capability does not have
its reset value, as this may be the consequence of the firmware
tempering with it and will impact all capabilities derived in
userspace.

Signed-off-by: Kevin Brodsky <kevin.brodsky@arm.com>