arm64: morello: Context-switch and initialise PCC
Store PCC in pt_regs and save/restore it on kernel_entry/kernel_exit
like the capability GPRs.

Extend start_thread() to initialise PCC to the root capability.
Also allow the process to be started in the C64 ISA (instead of A64) by
setting the LSB of its entry point.

Signed-off-by: Kevin Brodsky <kevin.brodsky@arm.com>