arm64: morello: Handle capability faults
Handle the new DFSC values allocated for capability faults by
delivering SIGSEGV to the faulting thread. A new si_code value is
allocated for each type of capability fault.

The value of SEGV_CAPTAGERR is identical to SEGV_CPERR. This is for
historical reasons (the SEGV_CAP* codes were allocated before
SEGV_CPERR appeared); a uapi change was not deemed justified as the
two codes are easy to disambiguate (one is x86/CET-only, the other
Morello/CHERI-only).

Signed-off-by: Kevin Brodsky <kevin.brodsky@arm.com>