arm64: morello: ptrace: Provide a setter for the Morello regset
Enable Morello registers to be set via ptrace by adding a .set
method to the NT_ARM_MORELLO regset, accessible using:

  ptrace(PTRACE_SETREGSET, pid, NT_ARM_MORELLO, &iov);

This operation is only available if the cheri.ptrace_forge_cap
sysctl is enabled (disabled by default). This is because it allows
injecting arbitrary capabilities in the tracee, bypassing any
capability restriction.

To prevent creating bogus capabilities with the tag set in the
tracee (which would also rely on the firmware enabling the SCTAG
instruction), if setting a capability with the tag set is requested,
the capability is derived from the root capability (using the BUILD
instruction).

Signed-off-by: Kevin Brodsky <kevin.brodsky@arm.com>