kselftests/arm64: Add basic Morello syscall tests
Add initial tests for purecap syscalls. The assembly implementation
is temporary and will be replaced with a C implementation later on.

Since only Clang supports Morello currently, it is used to build
the tests. They may be compiled from the top directory using:

  make -C tools/testing/selftests CLANG=clang TARGETS=arm64 ARM64_SUBTARGETS=morello ARCH=arm64

or just "make" from the current directory.

Signed-off-by: Amit Daniel Kachhap <amit.kachhap@arm.com>
Signed-off-by: Kevin Brodsky <kevin.brodsky@arm.com>