arm64: morello: Build the kernel in hybrid-capability mode
Require the compiler to support -march=morello and pass it for
building C, in addition to assembly. This turns the kernel into a
hybrid-capability binary; the kernel ABI remains mostly unchanged,
but capabilities can directly be manipulated in C.

Since we now require a Morello-aware compiler, there is no point in
keeping a special path for the case where a non-integrated assembler
is used and the compiler itself does not support Morello. Just pass
-march=morello both when assembling and compiling.

Signed-off-by: Kevin Brodsky <kevin.brodsky@arm.com>