include/linux/compat.h: Modify __SC_DELOUSE for 64 bit compat
__SC_DELOUSE typecasts all the received syscall parameter to
unsigned long. However this is unsuitable for PCuABI as the largest
parameter type is user_uintptr_t (__uintcap_t).

To address that the __SC_DELOUSE macro gets modified to, based on the type
of a syscall argument, eiter upcast it to a user_uintptr_t or create a valid
capability for user pointers.

Signed-off-by: Amit Daniel Kachhap <amit.kachhap@arm.com>
Co-developed-by: Beata Michalska <beata.michalska@arm.com>
Signed-off-by: Beata Michalska <beata.michalska@arm.com>