fs/aio: Fix user pointer downcasts
Do not cast a user pointer directly to a type that cannot hold a
CHERI capability, use user_ptr_addr() to extract its address.

These downcasts should be harmless because the address of the user
iocb struct is used as a sort of identifier, and not to access user
memory.

Signed-off-by: Kevin Brodsky <kevin.brodsky@arm.com>