fs/ioctl: Create a compat_noptr_ioctl helper
In some usecases, struct file_operations members: unlocked_ioctl and
compat_ioctl are assigned the same function. However with
CONFIG_CHERI_PURECAP_UABI enabled, function prototypes for both diverge,
with a different type being specified for the last argument.
This wrapper provides a remedy for that yielding a common interface for compat
ioctl handler that do expect integer (arithmetic) values as an ioctl parameter,
instead of user pointers. As such, the wrapper promotes the provided value
to appropriate type required by the unlocked_ioctl counterpart, issuing
a null-derived capability in case of the PCuABI.

Signed-off-by: Amit Daniel Kachhap <amit.kachhap@arm.com>
Co-developed-by: Beata Michalska <beata.michalska@arm.com>
Signed-off-by: Beata Michalska <beata.michalska@arm.com>