arm64: syscall: define system call table for 64-bit compat tasks
Define a system call table for 64-bit compat tasks. System call numbers
are identical in the standard AArch64 ABI and PCuABI, so we directly use
the generic unistd.h header for 64-bit compat.

Regarding choice of system call handler:
 1) If there is a compat handler, then we want to use that, because
    usually the compat handler exists to call compat_ptr() on any
    pointers inside structs passed as arguments. This is what we want
    for 64-bit compat, as we build capabilities in compat_ptr() from
    64-bit pointers in structs. So #define __SYSCALL_COMPAT, in order to
    select the compat handler.
 2) If there is a single handler shared by native and compat tasks, then
    we should just be able to use that, as there should be no pointers
    inside argument structs.

Note that all of these system call handlers expect any pointer arguments
(not in structs) to be capabilities, but 64-bit compat tasks do not pass
capabilities. This will be handled for the shared handler case in a
later patch using a system call wrapper. (For compat handlers it is
already handled in the COMPAT_SYSCALL_DEFINEx wrapper.)

There are a few exceptions where unistd.h chooses a handler not suitable
for 64-bit compat tasks. In this patch, we fix up the cases where
compilation fails (because the handler assigned by unistd.h does not
exist). Other cases will be handled in later patches as they are found.

Finally, also fix up the personality() system call, the same way it is
done on a non-purecap AArch64 system.

Signed-off-by: Kristina Martsenko <kristina.martsenko@arm.com>
Signed-off-by: Kevin Brodsky <kevin.brodsky@arm.com>