arm64: morello: Implement raw_copy_{from,to}_user_with_captags()
This patch implements the low-level tag-preserving uaccess routines.
Most of the existing logic for __arch_copy_{from,to}_user() is reused to
implement the new __arch_copy_{from,to}_user_with_captags().

copy_template.S is extended to copy capability tags for the
*_with_captags() variants (controlled by the COPY_CAPTAGS macro).
The approach is the same as a generic tag-preserving memcpy()
implementation (i.e. copy tags whenever possible, which requires at
least the source and destination addresses to be mutually aligned).

The existing {ldr,str}1 assembler macros can be reused as-is to
operate on capabilities, unfortunately this is not the case of
{ldp,stp}1, because they need to know the register size (8 for X, 16
for C). user_{ldp,stp} are extended and a new {ldp,stp}c1 pair
introduced accordingly.

Note: the implementation of __arch_copy_{from,to}_user() remains
unchanged (i.e. they never preserve tags).

Signed-off-by: Kevin Brodsky <kevin.brodsky@arm.com>