kselftests/arm64: Test the process environment
Add tests for the process environment so that a libc (e.g. musl) can
successfully start and call main. This includes validating the stack and
everything on it, checking for C64 state, write() and set_gettid().

This is necessary preparation to run the test harness instead of the
minimal freestanding environment.

Signed-off-by: Boyan Karatotev <boyan.karatotev@arm.com>
Signed-off-by: Kevin Brodsky <kevin.brodsky@arm.com>