linux/syscalls.h: Amend __SC_TEST to accommodate for capabilities
With the support for capabilities, the __SC_TEST macro assumptions
no longer stand when enabling PCuABI, due to increased size of user pointers.
Amend that by making sure the capabilities are properly recognized.
The verification here completely relies on compiler's CHERI intrinsics,
which are expected to throw an error when not dealing with an actual
capability.

Signed-off-by: Amit Daniel Kachhap <amit.kachhap@arm.com>
Co-developed-by: Beata Michalska <beata.michalska@arm.com>
Signed-off-by: Beata Michalska <beata.michalska@arm.com>
Signed-off-by: Kevin Brodsky <kevin.brodsky@arm.com>