linux/syscalls.h: Use uintcap_t type for syscall parameter in SC_LONG
When CONFIG_CHERI_PURECAP_UABI is used then the argument received as the
syscall parameter is of maximum 128 bit size and fits the type uintcap_t.

Hence receive them as type uintcap_t initially in the syscall wrapper
function __se_sys_{*}() instead of long or long long. Later they can be
typecasted to the actual type with __SC_CAST macro and passed to the final
syscall handler.

Signed-off-by: Amit Daniel Kachhap <amit.kachhap@arm.com>
Signed-off-by: Kevin Brodsky <kevin.brodsky@arm.com>