Makefile: Disable warning when using prefix __capability in PCuABI
We need to use the deprecated position for __capability, i.e. as a
prefix for *, as this is where __user is already positioned. This
causes the compiler to complain, along the lines of:

    warning: use of __capability before the pointer type is deprecated

Work around by that by passing -Wno-deprecated-declarations.
Unfortunately there isn't a more specific flag to disable just this
warning at the moment.

Signed-off-by: Kevin Brodsky <kevin.brodsky@arm.com>