arm64: morello: Remove unnecessary capability helpers
Now that capabilities are represented as uintcap_t, we can rely on
the compiler to perform appropriate capability operations, such as
assignment or address extraction. This allows us to simplify
existing code a bit by removing helpers that are now unnecessary.

Since capabilities are no longer represented using an opaque type,
there is no reason to pass read-only capability arguments by address
any more. Just pass them by value, which also saves us a load
instruction. Similarly, output capabilities can be returned
directly.

This patch is not comprehensive, ideally the vast majority of
functions in morello.S should be reimplemented in C for better
readability. This is left as a future improvement, a note to that
effect is added to morello.S.

Signed-off-by: Kevin Brodsky <kevin.brodsky@arm.com>