arm64: morello: add networking options to transitional config
Add a few additional options to the Morello transitional PCuABI
defconfig to enable networking.

CONFIG_PACKET is needed for DHCP.

CONFIG_SMC91X is needed to use the other network device on the FVP.

CONFIG_HW_RANDOM and CONFIG_HW_RANDOM_VIRTIO are needed to initialize
the kernel's random number pool on the FVP, which in turn is needed for
the SSH daemon to start up.

Signed-off-by: Kristina Martsenko <kristina.martsenko@arm.com>
Signed-off-by: Kevin Brodsky <kevin.brodsky@arm.com>